Game: get as close to 60fps as possible
This commit makes the timer wait until 1/60 seconds has passed before starting the next update or draw. It also calculates the time since the last frame, and passes that to each sprite's update method, which makes the game feel more smooth even if we're not at the target of 60fps.